Started implementing specialisation wizard
This commit is far from functional (well, it compiles, but that's about
it...). Plenty of ugly test/hacks/debug code. But there is a rough
outline for the first two screens of the wizard and I'm wrapping my
head around the GWT-RPC calls that will be needed to use or adapt to
support the wizard.