Massive refactoring of specialization wizard
Work is still ongoing. Core changes:

1) Each screen of wizard is now its own class for the sake of
modularity and simplicity.
2) Instead of piggybacking on top of existing actions/handlers/results
and coordinating a whole lot of service calls, we now make a single
service call to implement a specialization.
3) The UI is being updated to support the workflow in Ylva Hecktors MSc
thesis, with an added screen for specialization composition suggestions.
4) XDP-specific frames have been implemented to support the above,
rather than reuse attributes on treenodes and making the semantics of
things implicit in a variety of places its now all explicit and clear.